Serialize processing of secondary candidates.
Secondary cw and pulse candidates could be processed simultaneously,
when we prefer to do them sequentially.  Added a secondary candidate
list and a semaphore to control processing.  When a secondary
candidate message is received, it will be processed immediately
if no secondary candidate is currently being processed, but will
be added to a list if there is already a candidate in processing.
When processing is completed for a candidate, the next candidate
on the list is processed.